requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
roboflow>=1.1.0
Pillow>=9.0.0
numpy>=1.24.0
pycocotools>=2.0.7
tqdm>=4.65.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
